% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_cor.R
\name{prot_cor}
\alias{prot_cor}
\title{Perform correlation analysis using proteomics data file}
\usage{
prot_cor(
  samples,
  prot_quant_data,
  result_dir,
  continuous_var,
  method = "spearman"
)
}
\arguments{
\item{samples}{a data frame of sample information, including "Sample_ID", "Group" and covariates}

\item{prot_quant_data}{a data frame of proteomics quantification matrix with rows as genes and columns as samples}

\item{result_dir}{absolute path where the results should be stored}

\item{continuous_var}{continuous variable (this must be included in the sample_file)}

\item{method}{correlation analysis method, "spearman" or "pearson", Default="spearman"}
}
\value{
a data frame of correlation analysis result
}
\description{
`prot_cor()` conducts correlation analysis using spearman, pearson or kendall
}
\examples{
sample_file <- system.file("extdata", "apollo1.sample.info.tsv", package = "rpMeta")
rna_count_file <- system.file("extdata", "apollo1.rna.count.top6k.tsv", package = "rpMeta")
protein_quant_file <- system.file("extdata", "apollo1.protein.quant.top3k.tsv", package = "rpMeta")
inputs <- loadRpInput(sample_file, rna_count_file, protein_quant_file, format="tsv")
samples <- inputs$samples
rna_count_data <- inputs$rna
prot_quant_data <- inputs$protein
continuous_var <- "PkYrs"
method <- "spearman"
# prot_cor(samples, prot_quant_data, result_dir, continuous_var=continuous_var, method=method)
}
