% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_de.R
\name{prot_de}
\alias{prot_de}
\title{Calculate differrential expression from proteomics data file}
\usage{
prot_de(
  samples,
  prot_quant_data,
  result_dir,
  group1 = "TRU",
  group2 = "ProxProlif",
  covariates = NULL,
  method = "limma",
  X11 = "No",
  ...
)
}
\arguments{
\item{samples}{a data frame of sample information, including "Sample_ID", "Group" and covariates}

\item{prot_quant_data}{a data frame of proteomics quantification matrix with rows as genes and columns as samples}

\item{result_dir}{absolute path where the results should be stored}

\item{group1}{name of the first group used in DE analysis, a.k.a condition1 or factor1}

\item{group2}{name of the second group used in DE analysis, a.k.a condition2 or factor2
fold change is calculated as group2/group1}

\item{covariates}{covariates used in the model (this must be included in the sample_file)}

\item{method}{Differential expression analysis method, Default="limma"}

\item{X11}{output X11 or PDF for graphics, c("No", "Yes"), (Default="No")}

\item{...}{Additional arguments 
`limma_proportion` numeric value between 0 and 1, assumed proportion of genes which are differentially expressed}
}
\value{
a data frame of DE anlaysis result
}
\description{
`prot_de()` conducts differential expression analysis using limma. 
Covariates can be added to the model.
}
\examples{
sample_file <- system.file("extdata", "apollo1.sample.info.tsv", package = "rpMeta")
rna_count_file <- system.file("extdata", "apollo1.rna.count.top6k.tsv", package = "rpMeta")
protein_quant_file <- system.file("extdata", "apollo1.protein.quant.top3k.tsv", package = "rpMeta")
inputs <- loadRpInput(sample_file, rna_count_file, protein_quant_file, format="tsv")
samples <- inputs$samples
rna_count_data <- inputs$rna
prot_quant_data <- inputs$protein
group1 <- "TRU"
group2 <- "ProxProlif"
covariates <- c("Sex", "grade")
prot_method <- "limma"
result_dir <- file.path(getwd(), "rpMeta_APOLLO1")
limma_proportion <- 0.02
if (!dir.exists(result_dir)) { dir.create(result_dir) }
prot_de(samples, prot_quant_data, result_dir, group1, group2, covariates, method=prot_method, X11="Yes", limma_proportion=limma_proportion)
}
