% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_de.R
\name{rna_de}
\alias{rna_de}
\title{Perform differrential expression analysis using RNA count matrix}
\usage{
rna_de(
  samples,
  rna_count_data,
  result_dir,
  group1,
  group2,
  method = "limma",
  covariates = NULL,
  normalization = "default",
  X11 = "No",
  ...
)
}
\arguments{
\item{samples}{a data frame of sample information, including "Sample_ID", "Group" and covariates}

\item{rna_count_data}{a data frame of RNASeq raw count matrix with rows as genes and columns as samples}

\item{result_dir}{absolute path where the results should be stored}

\item{group1}{name of the first group used in DE analysis, a.k.a condition1 or factor1}

\item{group2}{name of the second group used in DE analysis, a.k.a condition2 or factor2
fold change is calculated as group2/group1}

\item{method}{DE analysis method, limma, DESeq2 or user defined function, Default="limma"}

\item{covariates}{covariates used in the model (this must be included in the sample_file)}

\item{normalization}{normalization method, Default="default"}

\item{X11}{output X11 or PDF for graphics, c("No", "Yes"), (Default="No")}

\item{...}{Additional arguments. If a named argument `mean_rna_count` (mean RNA count cutoff per row/gene) is supplied (e.g. 5)}
}
\value{
a data frame of DE anlaysis result
}
\description{
`rna_de()` conducts differential expression analysis using limma or DESeq2.
Covariates can be added to the model
}
\examples{
sample_file <- system.file("extdata", "apollo1.sample.info.tsv", package = "rpMeta")
rna_count_file <- system.file("extdata", "apollo1.rna.count.top6k.tsv", package = "rpMeta")
protein_quant_file <- system.file("extdata", "apollo1.protein.quant.top3k.tsv", package = "rpMeta")
inputs <- loadRpInput(sample_file, rna_count_file, protein_quant_file, format="tsv")
samples <- inputs$samples
rna_count_data <- inputs$rna
prot_quant_data <- inputs$protein
group1 <- "TRU"
group2 <- "ProxProlif"
covariates <- c("Sex", "grade")
rna_method <- "limma"
rna_norm <- "default"
result_dir <- file.path(getwd(), "rpMeta_example")
mean_rna_count <- 5 
DESeq2_test <- "Wald" 
if (!dir.exists(result_dir)) { dir.create(result_dir) }
rna_de(samples, rna_count_data, result_dir, group1, group2, method=rna_method, covariates, rna_norm, X11="Yes", mean_rna_count=mean_rna_count, DESeq2_test = DESeq2_test)
}
