% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_de.R
\name{rna_normalize}
\alias{rna_normalize}
\title{Perform RNASeq count matrix normalization using log2(CPM) and median center}
\usage{
rna_normalize(count_data)
}
\arguments{
\item{count_data}{RNASeq count data frame, rows are genes and columns are samples.}
}
\value{
a data frame of normalized RNA quantification matrix
}
\description{
`rna_normalize()` conducts RNASeq count matrix normalization
}
