% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpCorGenePlot.R
\name{rpCorGenePlot}
\alias{rpCorGenePlot}
\title{`rpCorGenePlot()` plot RNA–protein scatterplot and boxplot for top 20 DEGs}
\usage{
rpCorGenePlot(
  df,
  rna_norm_counts,
  prot_norm_counts,
  samples,
  continuous_var,
  result_dir = "./",
  genes = 20,
  X11 = "No"
)
}
\arguments{
\item{df}{combined RNA–protein differential analysis result data frame}

\item{rna_norm_counts}{an RNA count data frame with samples as columns and gene symbols as rows}

\item{prot_norm_counts}{an proteomics count data frame with samples as columns and gene symbols as rows}

\item{samples}{sample description data frame with group and covariates columns}

\item{continuous_var}{name of the continuous variable}

\item{result_dir}{absolute path where the results should be stored, default="./"}

\item{genes}{a vector of gene symbols, e.g. c("MET","BRCA2") or top n by Meta_FDR (default=20)}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}
}
\value{
A data.frame with columns \code{gene_name}, \code{RNA_pvalue}, \code{Prot_pvalue}, \code{Meta_pvalue}, \code{Spearman_rho}, \code{Spearman_pvalue}
}
\description{
`rpCorGenePlot()` plot RNA–protein scatterplot and boxplot for top 20 DEGs
}
\examples{
# rpCorGenePlot(df=m_cor, rna_norm_counts=rna_norm_counts, prot_norm_counts=prot_norm_counts,
#            samples=samples, continuous_var, result_dir)
}
