% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpCorHeatmap.R
\name{rpCorHeatmap}
\alias{rpCorHeatmap}
\title{Plot RNA-protein heatmap for continuous variable correlation analysis}
\usage{
rpCorHeatmap(
  rpCor,
  rna_norm_counts,
  prot_norm_counts,
  samples,
  continuous_var,
  result_dir = "./",
  rpscale = "Yes",
  boundN = 2,
  genesToDisplay = 20,
  X11 = "No",
  ...
)
}
\arguments{
\item{rpCor}{combined RNA protein correlation analysis result data frame}

\item{rna_norm_counts}{an RNA count data frame with samples as columns and gene symbols as rows}

\item{prot_norm_counts}{a proteomics quantification data frame with samples as columns and gene symbols as rows}

\item{samples}{sample description data frame with group and covariates columns}

\item{continuous_var}{continuous variable used in the analysis}

\item{result_dir}{absolute path where the results should be stored, default="./"}

\item{rpscale}{Z score normalization or not (default="Yes")}

\item{boundN}{minimum and maximum value of the heatmap color scale}

\item{genesToDisplay}{Either a numeric value indicating the top n meta DEGs (by Meta_PVal) to display or a vector/list of gene symbols.}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}

\item{...}{Additional arguments. 
`legend` True or False, legend to be drawn or not  
`scale` character indicating if the values should be centered and
         scaled in either the row direction or the column direction,
         or none. Corresponding values are ‘"row"’, ‘"column"’ and ‘"none"’}
}
\value{
None
}
\description{
`rpCorHeatmap()` plots an integrated heatmap for joint RNA and protein expression visualization (for continuous variable).
rpHeatmap.cor.meta.pdf displays RNA and protein expression side by side 
heatmap.cor.pdf displays RNA and protein expression separately
}
\examples{
# rpCorHeatmap(m, rna_norm_counts, prot_norm_counts, samples, continuous_var=continuous_var, result_dir, genesToDisplay=20)
}
