% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpCorPathwayPlot.R
\name{rpCorPathwayPlot}
\alias{rpCorPathwayPlot}
\title{Create pathway enrichment plots for correlation analysis}
\usage{
rpCorPathwayPlot(
  df,
  result_dir,
  continuous_var,
  pathways = "Hallmark",
  FDR_cutoff = 0.05,
  max_pathways = 30,
  X11 = "No"
)
}
\arguments{
\item{df}{pathway enrichment result data frame}

\item{result_dir}{Absolute path where the results should be stored.}

\item{continuous_var}{continuous variable used in the analysis}

\item{pathways}{Pathway database (Hallmark, KEGG, Cell_Type_Signature, IMMUNESIGDB or GO_BP). Default="Hallmark"}

\item{FDR_cutoff}{false discovery rate (FDR) cutoff for plot, default=0.05}

\item{max_pathways}{maximum number of pathways to be displayed in the plot, default=30}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}
}
\value{
None
}
\description{
`rpCorPathwayPlot()` create a combined RNA/protein/meta pathway enrichment analysis bubble plot for correlaiton analysis.
}
