% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpGenePlot.R
\name{rpGenePlot}
\alias{rpGenePlot}
\title{`rpGenePlot()` plot RNA–protein scatterplot and boxplot for top 20 DEGs}
\usage{
rpGenePlot(
  df,
  rna_norm_counts,
  prot_norm_counts,
  samples,
  group1,
  group2,
  result_dir = "./",
  genes = 20,
  X11 = "No"
)
}
\arguments{
\item{df}{combined RNA–protein differential analysis result data frame}

\item{rna_norm_counts}{an RNA count data frame with samples as columns and gene symbols as rows}

\item{prot_norm_counts}{an proteomics count data frame with samples as columns and gene symbols as rows}

\item{samples}{sample description data frame with group and covariates columns}

\item{group1}{name of the first group used in DE analysis}

\item{group2}{name of the second group used in DE analysis}

\item{result_dir}{absolute path where the results should be stored, default="./"}

\item{genes}{a vector of gene symbols, e.g. c("MET","BRCA2") or top n by Meta_FDR (default=20)}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}
}
\value{
A data.frame with columns \code{gene_name}, \code{RNA_pvalue}, \code{Prot_pvalue}, \code{Meta_pvalue}, \code{Spearman_rho}, \code{Spearman_pvalue}
}
\description{
`rpGenePlot()` plot RNA–protein scatterplot and boxplot for top 20 DEGs
}
\examples{
deg_file <- system.file("extdata","DEG.combined.txt",package="rpMeta")
m <- read.table(deg_file,header=TRUE,stringsAsFactors=FALSE)
# rpGenePlot(df=m, rna_norm_counts=rna_norm_counts, prot_norm_counts=prot_norm_counts,
#            samples=samples, group1, group2, result_dir)
}
