% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpHeatmap.R
\name{rpHeatmap}
\alias{rpHeatmap}
\title{Plot RNA-protein heatmap plot}
\usage{
rpHeatmap(
  rpDEG,
  rna_norm_counts,
  prot_norm_counts,
  samples,
  covariates = NULL,
  result_dir = "./",
  rpscale = "Yes",
  boundN = 2,
  genesToDisplay = 20,
  X11 = "No",
  rpColCluster = "group",
  ...
)
}
\arguments{
\item{rpDEG}{combined RNA protein differential analysis result data frame}

\item{rna_norm_counts}{an RNA count data frame with samples as columns and gene symbols as rows}

\item{prot_norm_counts}{a proteomics quantification data frame with samples as columns and gene symbols as rows}

\item{samples}{sample description data frame with group and covariates columns}

\item{covariates}{covariates used in the model, Default=NULL}

\item{result_dir}{absolute path where the results should be stored, default="./"}

\item{rpscale}{Z score normalization or not (default="Yes")}

\item{boundN}{minimum and maximum value of the heatmap color scale}

\item{genesToDisplay}{Either a numeric value indicating the top n meta DEGs (by Meta_PVal) to display or a vector/list of gene symbols.}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}

\item{rpColCluster}{option to cluster colums(samples), "group" or "all", default="group"}

\item{...}{Additional arguments. 
`legend` True or False, legend to be drawn or not  
`scale` character indicating if the values should be centered and
         scaled in either the row direction or the column direction,
         or none. Corresponding values are ‘"row"’, ‘"column"’ and ‘"none"’}
}
\value{
None
}
\description{
`rpHeatmap()` plots an integrated heatmap for joint RNA and protein expression visualization.
rpMeta.rpHeatmap.meta.[timestamp].pdf displays RNA and protein expression side by side
samples are clustered within each group by RNA expression 
rpMeta.heatmap.DEG.[timestamp].pdf displays RNA and protein expression separately
}
\examples{
# rpHeatmap(m, rna_norm_counts, prot_norm_counts, samples, covariates=covariates, result_dir, genesToDisplay=20)
}
