% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpPathway.R
\name{rpPathway}
\alias{rpPathway}
\title{Perform pathway enrichment analysis}
\usage{
rpPathway(
  df,
  result_dir,
  pathways = "Hallmark",
  FDR_cutoff = 0.25,
  species = "Homo sapiens"
)
}
\arguments{
\item{df}{Combined RNA-protein differential analysis result data frame.}

\item{result_dir}{Absolute path where the results should be stored.}

\item{pathways}{Pathway database (Hallmark, KEGG, Cell_Type_Signature, IMMUNESIGDB or GO_BP), default="Hallmark"}

\item{FDR_cutoff}{false discovery rate (FDR) cutoff, default=0.25}

\item{species}{Species, e.g., "Homo sapiens".}
}
\value{
a data frame of pathway enrichment results
}
\description{
`rpPathway()` performs KEGG, Cell_Type_Signature, IMMUNESIGDB, GO:BP, 
or Hallmark pathway enrichment analysis
Analysis is based on GSEA method using fgsea R package
Molecular Signatures Database (MSigDB) gene sets are used through msigdbr R package 
Subramanian A et al. Gene set enrichment analysis: a knowledge-based approach for 
interpreting genome-wide expression profiles. Proc Natl Acad Sci U S A. 2005 Oct 25;102(43)
}
