% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpPathwayPlot.R
\name{rpPathwayPlot}
\alias{rpPathwayPlot}
\title{Create pathway enrichment plots}
\usage{
rpPathwayPlot(
  df,
  result_dir,
  group1,
  group2,
  pathways = "Hallmark",
  FDR_cutoff = 0.05,
  max_pathways = 30,
  X11 = "No"
)
}
\arguments{
\item{df}{Combined RNA-protein differential analysis result data frame.}

\item{result_dir}{Absolute path where the results should be stored.}

\item{group1}{name of the first group used in DE analysis}

\item{group2}{name of the second group used in DE analysis}

\item{pathways}{Pathway database (Hallmark, KEGG, Cell_Type_Signature, IMMUNESIGDB or GO_BP). Default="Hallmark"}

\item{FDR_cutoff}{false discovery rate (FDR) cutoff for plot, default=0.05}

\item{max_pathways}{maximum number of pathways to be displayed in the plot, default=30}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}
}
\value{
None
}
\description{
`rpPathwayPlot()` create a combined RNA/protein/meta pathway enrichment analysis bubble plot.
}
