% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpVenn.R
\name{rpVenn}
\alias{rpVenn}
\title{Plot Venn diagram of RNA, protein and meta DEGs}
\usage{
rpVenn(
  rpDEG,
  fdr_cutoff = 0.05,
  result_dir = "./",
  filename_suffix = "",
  X11 = "No"
)
}
\arguments{
\item{rpDEG}{combined RNA protein differential analysis result data frame}

\item{fdr_cutoff}{The FDR threshold RNA, protein and meta DEGs in rpVenn plot}

\item{result_dir}{absolute path where the results should be stored, default="./"}

\item{filename_suffix}{file name suffix of the output pdf file}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}
}
\value{
A list of counts, RNA, protein, meta, two way overlaps, three way overlaps
}
\description{
`rpVenn` Plot Venn diagram of RNA, protein and meta DEGs
}
\examples{
deg_file <- system.file("extdata", "DEG.combined.txt", package = "rpMeta")
rpDEG <- read.table(deg_file, header=TRUE, stringsAsFactors=FALSE)
result_dir <- file.path(getwd(), "rpMeta_APOLLO1")
if (!dir.exists(result_dir)) { dir.create(result_dir) }
rpVenn(rpDEG, fdr_cutoff=0.05, result_dir)
}
