% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpVolcano.R
\name{rpVolcano}
\alias{rpVolcano}
\title{Plot RNA-protein volcano plot}
\usage{
rpVolcano(
  rpDEG,
  result_dir,
  group1,
  group2,
  FDR_cutoff = 0.05,
  log2FC_cutoff = 1,
  genesToDisplay = 5,
  X11 = "No"
)
}
\arguments{
\item{rpDEG}{Combined RNA protein differential analysis result data frame.}

\item{result_dir}{Absolute path where the results should be stored.}

\item{group1}{name of the first group used in DE analysis}

\item{group2}{name of the second group used in DE analysis}

\item{FDR_cutoff}{FDR cutoff in volcano plot, Default = 0.05.}

\item{log2FC_cutoff}{log2 Fold Change cutoff in volcano plot, Default = 1 (FoldChange = 2).}

\item{genesToDisplay}{a numeric value (default 5) indicating the top n up‐ and down‐regulated genes (based on Meta_FDR) to annotate.}

\item{X11}{"No" (default) to write PDF, or "Yes" to plot to X11 device}
}
\value{
Null
}
\description{
`rpVolcano()` plots RNA-protein volcano plots and fold change.
}
\examples{
deg_file <- system.file("extdata", "DEG.combined.txt", package = "rpMeta")
m <- read.table(deg_file, header = TRUE, stringsAsFactors = FALSE)
result_dir <- file.path(getwd(), "rpMeta_APOLLO1")
group1 <- "TRU"
group2 <- "ProxProlif"
if (!dir.exists(result_dir)) { dir.create(result_dir) }
rpVolcano(m, result_dir, group1, group2, genesToDisplay = 5, X11="Yes")
}
